+++ routes/queryPublicPhonesRoutes.js
@@ -2,12 +2,20 @@ const { Router } = require('express');
 const routes = require('.');
 const connection = require('../app/db.config');
 const { createValidPublicPhone, updateValidPublicPhone } = require('../middlewares/public_phones.validation.middleware');
+const { authenticateJWT } = require('../middlewares/authorization.middleware');
 
 const router = Router();
 
 // add a new public phone
-router.post('/', (req, res, next) => { 
+router.post('/', authenticateJWT, (req, res, next) => { 
     try {
+        const { role } = req.user;
+        if (role !== 'admin') {
+            return res.status(403).send({
+               error: true,
+               message: "Forbidden"
+            });
+       }
         const publicPhoneData = {
             public_phone_type: req.body.public_phone_type,
             station_id: req.body.station_id,
@@ -41,8 +49,15 @@ router.post('/', (req, res, next) => {
 });
 
 // update public phone
-router.put('/:public_phone_id', async (req, res, next) => { 
+router.put('/:public_phone_id', authenticateJWT, async (req, res, next) => { 
     try {
+        const { role } = req.user;
+        if (role !== 'admin') {
+            return res.status(403).send({
+               error: true,
+               message: "Forbidden"
+            });
+       }
         const updatedData = {
             public_phone_type: req.body.public_phone_type,
             station_id: req.body.station_id,
@@ -69,8 +84,15 @@ router.put('/:public_phone_id', async (req, res, next) => {
 });
 
 // delete a public phone
-router.delete('/:public_phone_id', (req, res, next) => {
+router.delete('/:public_phone_id', authenticateJWT, (req, res, next) => {
     try {
+        const { role } = req.user;
+        if (role !== 'admin') {
+            return res.status(403).send({
+               error: true,
+               message: "Forbidden"
+            });
+       }
         connection.query(`DELETE FROM public_phones WHERE public_phone_id = ${req.params.public_phone_id}`, (error, results, fields) => {
             if(error){
                 throw error;
