+++ routes/queryStationsRoutes.js
@@ -2,12 +2,20 @@ const { Router } = require('express');
 const routes = require('.');
 const connection = require('../app/db.config');
 const { createValidStation, updateValidStation } = require('../middlewares/stations.validation.middleware');
+const { authenticateJWT } = require('../middlewares/authorization.middleware');
 
 const router = Router();
 
 // add a new station
-router.post('/', (req, res, next) => { 
+router.post('/', authenticateJWT, (req, res, next) => { 
     try {
+        const { role } = req.user;
+        if (role !== 'admin') {
+            return res.status(403).send({
+               error: true,
+               message: "Forbidden"
+            });
+       }
         const station_data = {
             station_id: req.body.station_id,
             station_type: req.body.station_type
@@ -35,8 +43,15 @@ router.post('/', (req, res, next) => {
 });
 
 // update station
-router.put('/:station_id', async (req, res, next) => { 
+router.put('/:station_id', authenticateJWT, async (req, res, next) => { 
     try {
+        const { role } = req.user;
+        if (role !== 'admin') {
+            return res.status(403).send({
+               error: true,
+               message: "Forbidden"
+            });
+       }
         const updatedData = {
             station_type: req.body.station_type
         };
@@ -48,7 +63,6 @@ router.put('/:station_id', async (req, res, next) => {
         next();
     }
 }, updateValidStation, (req, res, next) => {
-    console.log(res.data.station_type, req.params.station_id);
     connection.query(`UPDATE stations 
         SET station_type = '${res.data.station_type}'
         WHERE station_id = ${req.params.station_id}`, (error, results, fields) => {
@@ -57,8 +71,15 @@ router.put('/:station_id', async (req, res, next) => {
 });
 
 // delete a station
-router.delete('/:station_id', (req, res, next) => {
+router.delete('/:station_id', authenticateJWT, (req, res, next) => {
     try {
+        const { role } = req.user;
+        if (role !== 'admin') {
+            return res.status(403).send({
+               error: true,
+               message: "Forbidden"
+            });
+       }
         connection.query(`DELETE FROM stations WHERE station_id = ${req.params.station_id}`, (error, results, fields) => {
             if(error){
                 throw error;
